requirements.txt
ray[data]>=2.44.1
transformers>=4.36.0
accelerate>=0.25.0
torch>=2.1.0
huggingface-hub>=0.19.0
fsspec>=2023.9.0
Pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0